Customize generated Go code
The `msgp` library doesn't cater to structs with int keys. So we
generate the code assuming string keys and then change the code relevant
to int key decoding. If you need to see a difference between the
generated code and the customized code, run `go generate hw.go` which
will update the file `hw_gen_new.go`.

This is pretty complicated if there are a lot of types since we'd have
to ensure that we change the items everywhere. A better way will be to
use msgpack tuples or simply string keys. Not to mention the test
generated by the code generator won't run.